+++ src/UserService/Domain/Repositories/IRepositoryBase.cs
@@ -1,6 +1,6 @@
 using Domain.Entities;
 
-namespace Infrastructure.Repository;
+namespace Domain.Repositories;
 
 public interface IRepositoryBase<TEntity> where TEntity : EntityBase
     {
