+++ src/UserService/Domain/Repositories/IUserRepository.cs
@@ -1,6 +1,6 @@
 using Domain.Entities;
 
-namespace Infrastructure.Repository;
+namespace Domain.Repositories;
 
 public interface IUserRepository : IRepositoryBase<User>
 {
