+++ src/UserService/Infrastructure/Security/PasswordHasher.cs
@@ -1,0 +1,14 @@
+namespace Infrastructure.Security;
+
+public class PasswordHasher
+{
+    public string HashPassword(string password)
+    {
+        throw new NotImplementedException();
+    }
+
+    public string VerifyPassword(string hashedPassword, string password)
+    {
+        throw new NotImplementedException();
+    }
+}
